vmm: net: Workaround OpenAPI generator limitations
OpenAPI generator, the tool used to generate the client code for Cloud
Hypervisor, has to deal with some specifics of the language the code is
generated for.

When we think about Kata Containers, we have to think (at least for now)
about golang, and golang has no knowledge about "None" and "Some",
leading to initialising structure fields as empty expressing those are
not set.  However, when this gets read on the cloud-hypervisor side it's
interpreted as something initialised and ready to be used (or, not
"None").

Considering everything said above, trying to use file descriptors
instead of the tap name would never work when cloud-hypervisor is used
with Kata Containers, as the tap name is never ever None, but rather an
empty string.

The best we can do, right now, is to actually take into account such
limitation and check whether the field is empty.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>